requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.1.2  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.0.1  # https://github.com/evansd/whitenoise
redis==3.5.0  # https://github.com/andymccurdy/redis-py
hiredis==1.0.1  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.11.0  # https://github.com/jazzband/django-redis
django-bootstrap3==12.1.0
django-simple-history==2.9.0

# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework

asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
ecdsa==0.13.3
future==0.18.2
html5lib==1.0.1
httplib2==0.11.3
idna==2.9
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.42
requests==2.23.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.25.9
webencodings==0.5.1
xhtml2pdf==0.2.4
